requirements.txt
transformers==4.41.2
torch>=2.0
beautifulsoup4==4.12.3
requests==2.32.3
googlesearch-python==1.2.3
python-pptx==0.6.23
matplotlib==3.8.4
schedule==1.2.1
pytrends==4.9.2
pillow==10.3.0
pandas==2.2.2
duckduckgo-search==3.9.5
readability-lxml==0.8.1
ratelimit==2.2.1
backoff==2.2.1
pytest==8.2.0
pytest-mock==3.14.0